models updated to handle high volume data
1. film_survey models: the source data now has inflated to result in the total number of views size beyond what postgres INT type can contain, hence they have been cast to bigint to be able to handle bigger values

2. beneficiary_case_data_agg: due to query running for too long, connection timeout was happening. query has been optimized now.

+++ models/prod/indirect_end_user/film_survey_film_metrics.sql
@@ -14,9 +14,9 @@ SELECT
     district_name,
     TRIM(individual_film_name) AS film_name,
     SUM(
-        COALESCE(men_viewers, 0)
-        + COALESCE(women_viewers, 0)
-        + COALESCE(other_viewers, 0)
+        COALESCE(men_viewers, 0)::BIGINT
+        + COALESCE(women_viewers, 0)::BIGINT
+        + COALESCE(other_viewers, 0)::BIGINT
     ) AS total_viewer_count
 FROM films_expanded
 WHERE TRIM(individual_film_name) != ''
